nav drawer partial to class
