+++ src/layouts/chat/index.js
@@ -1,20 +1,13 @@
 import template from './template.hbs'
 import './style.css'
-
 import nav from '../../config/nav'
 import icons from '../../config/icons'
-import { useNavDrawer } from '../../components/nav/nav-drawer'
 import Component from '../../lib/dom/Component'
+import NavDrawer from '../../components/ui/nav/nav-drawer'
 
 class ChatLayout extends Component {
   constructor(options) {
     super(options)
-  }
-
-  async mounted() {
-    super.mounted()
-
-    await useNavDrawer()
   }
 }
 
@@ -22,7 +15,11 @@ export default new ChatLayout({
   selector: '.chat-layout',
   template: template,
   props: {
-    navList: nav.drawer,
-    toggleIcon: icons.toggleNav
-  }
+    NavDrawer: NavDrawer.template({
+      ...NavDrawer.props,
+      navList: nav.drawer,
+      toggleIcon: icons.toggleNav
+    })
+  },
+  components: { NavDrawer }
 })
