+++ src/layouts/chat/template.hbs
@@ -1,3 +1,3 @@
-{{> nav-drawer }}
+{{{ NavDrawer }}}
 <div class="container full chat-layout">
 </div>
