+++ src/layouts/default/template.hbs
@@ -1,1 +1,1 @@
-{{> nav-drawer }}
+{{{ NavDrawer }}}
